requirements.txt
openai
PyPDF2
requests
urllib3
PyMuPDF
Pillow